Complete ctc dataset

+++ src/csj/labels/ctc/monolog/character.py
@@ -5,8 +5,8 @@
 
 import os
 import re
-import codecs
 import numpy as np
+import pandas as pd
 from tqdm import tqdm
 
 from prepare_path import Prepare
@@ -22,11 +22,11 @@ from .fix_trans import fix_transcript
 # = 145 + 1 + = 146 labels
 
 # [phone]
-#
-#
+# 39 phones, sil(_),
+# = 36 + 1 = 37 labels
 
 def read_sdb(label_paths, label_type, save_path=None):
-    """Read transcripts (.sdb) & save as npy files.
+    """Read transcripts (.sdb) & save files (.npy).
     Args:
         label_paths: list of paths to label files
         label_type: character or phone
@@ -34,75 +34,87 @@ def read_sdb(label_paths, label_type, save_path=None):
     Returns:
         speaker_dict: dictionary of speakers
             key => speaker name
-            value => dictionary of utterance infomation of each speaker
+            value => dictionary of utterance information of each speaker
                 key => utterance index
                 value => [start_frame, end_frame, transcript]
     """
     print('===> Reading target labels...')
     speaker_dict = {}
     char_set = set([])
-    for i, label_path in enumerate(tqdm(label_paths)):
+    for label_path in tqdm(label_paths):
+        col_names = [j for j in range(25)]
+        df = pd.read_csv(label_path, names=col_names,
+                         encoding='SHIFT-JIS', delimiter='\t', header=None)
+        drop_column = col_names
+        drop_column.remove(3)
+        drop_column.remove(5)
+        drop_column.remove(10)
+        drop_column.remove(11)
+        df = df.drop(drop_column, axis=1)
+
         utterance_dict = {}
         utt_index_pre = 1
-        start_frame_pre = None
+        start_frame_pre, end_frame_pre = None, None
         transcript = ''
-        with codecs.open(label_path, 'r', 'shift-jis') as f:
-            # print(label_path)
-            for line in f:
-                speaker_name = os.path.basename(label_path).split('.')[0]
-                line = line.strip().split('\t')
-                time_info = line[3].split(' ')
-                utt_index = int(time_info[0])
-                segment = time_info[1].split('-')
-                start_frame = int(float(segment[0]) * 100)
-                end_frame = int(float(segment[1]) * 100)
-                if start_frame_pre is None:
-                    start_frame_pre = start_frame
+        speaker_name = os.path.basename(label_path).split('.')[0]
+        for key, row in df.iterrows():
+            time_info = row[3].split(' ')
+            utt_index = int(time_info[0])
+            segment = time_info[1].split('-')
+            start_frame = int(float(segment[0]) * 100 + 0.5)
+            end_frame = int(float(segment[1]) * 100 + 0.5)
+            if start_frame_pre is None:
+                start_frame_pre = start_frame
+            if end_frame_pre is None:
+                end_frame_pre = end_frame
 
-                # word = line[5]  # include kanji characters
-                yomi = line[10]
-                # pos_tag = line[11] if len(line) >= 12 else None
+            word = row[5]  # include kanji characters
+            yomi = row[10]
+            pos_tag = row[11]
 
-                # stack word in the same utterance
-                if utt_index == utt_index_pre:
+            # stack word in the same utterance
+            if utt_index == utt_index_pre:
+                transcript += yomi
+                utt_index_pre = utt_index
+                end_frame_pre = end_frame
+                continue
+            else:
+                # count the number of kakko
+                left = transcript.count('(')
+                right = transcript.count(')')
+
+                if left != right:
                     transcript += yomi
                     utt_index_pre = utt_index
+                    end_frame_pre = end_frame
                     continue
                 else:
-                    # count the number of kakko
-                    left = transcript.count('(')
-                    right = transcript.count(')')
+                    # clean transcript
+                    transcript_fixed = fix_transcript(transcript)
 
-                    if left != right:
-                        transcript += yomi
-                        utt_index_pre = utt_index
-                        continue
-                    else:
-                        # clean transcript
-                        transcript_fixed = fix_transcript(transcript)
+                    # skip silence
+                    if transcript_fixed != '':
+                        # merge silence around each utterance
+                        transcript_fixed = '_' + transcript_fixed + '_'
 
-                        # skip silence
-                        if transcript_fixed != '':
-                            # merge silence around each utterance
-                            transcript_fixed = '_' + transcript_fixed + '_'
+                        # remove double underbar
+                        transcript_fixed = re.sub(
+                            '__', '_', transcript_fixed)
 
-                            # remove double underbar
-                            transcript_fixed = re.sub(
-                                '__', '_', transcript_fixed)
+                        for char in list(transcript_fixed):
+                            char_set.add(char)
 
-                            for char in list(transcript_fixed):
-                                char_set.add(char)
+                        utterance_dict[str(utt_index - 1).zfill(4)] = [
+                            start_frame_pre, end_frame_pre, transcript_fixed]
 
-                            utterance_dict[utt_index] = [
-                                start_frame_pre, end_frame, transcript_fixed]
+                    # initialization
+                    transcript = yomi
+                    utt_index_pre = utt_index
+                    start_frame_pre = start_frame
+                    end_frame_pre = end_frame
 
-                        # initialization
-                        transcript = yomi
-                        utt_index_pre = utt_index
-                        start_frame_pre = None
-
-            # register all utterances of each speaker
-            speaker_dict[speaker_name] = utterance_dict
+        # register all utterances of each speaker
+        speaker_dict[speaker_name] = utterance_dict
 
     # make mapping dictionary from kana to phone
     prep = Prepare()
@@ -126,17 +138,18 @@ def read_sdb(label_paths, label_type, save_path=None):
         file_name = 'char2num.txt'
     elif label_type == 'phone':
         file_name = 'phone2num.txt'
-    mapping_file_path = os.path.join(
-        prep.run_root_path, 'labels/ctc', file_name)
-    with open(mapping_file_path, 'w') as f:
-        if label_type == 'character':
-            for index, char in enumerate(kana_list):
-                f.write('%s  %s\n' % (char, str(index)))
-        elif label_type == 'phone':
-            for index, phone in enumerate(sorted(list(phone_set))):
-                if index == 0:
-                    f.write('_  0\n')
-                f.write('%s  %s\n' % (phone, str(index + 1)))
+    mapping_file_path = os.path.join(prep.run_root_path,
+                                     'labels/ctc', file_name)
+    if not os.path.isfile(mapping_file_path):
+        with open(mapping_file_path, 'w') as f:
+            if label_type == 'character':
+                for index, char in enumerate(kana_list):
+                    f.write('%s  %s\n' % (char, str(index)))
+            elif label_type == 'phone':
+                for index, phone in enumerate(sorted(list(phone_set))):
+                    if index == 0:
+                        f.write('_  0\n')
+                    f.write('%s  %s\n' % (phone, str(index + 1)))
 
     # for debug
     # for char in list(char_set):
@@ -150,7 +163,7 @@ def read_sdb(label_paths, label_type, save_path=None):
             mkdir(os.path.join(save_path, speaker_name))
             for utt_index, utt_info in utterance_dict.items():
                 start_frame, end_frame, transcript = utt_info
-                save_file_name = speaker_name + '_' + str(utt_index) + '.npy'
+                save_file_name = speaker_name + '_' + utt_index + '.npy'
 
                 if label_type == 'character':
                     # convert from kana character to number
